Only move files aside on Windows. And ensure that we stay in the same directory.
