+++ cmd/state-installer/internal/installer/installation_linux.go
@@ -1,8 +1,9 @@
-// +build linux windows
-
 package installer
 
 func InstallSystemFiles(_, _, _ string) error {
 	return nil
 }
 
+func (i *Installation) PrepareBinTargets() error {
+	return nil
+}
